Rename file and move docstring
